Add an option to not flush redis every time sidekiq.rb is loaded in development

+++ config/initializers/sidekiq.rb
@@ -2,7 +2,7 @@ require "#{Rails.root}/lib/discourse_redis"
 
 $redis = DiscourseRedis.new
 
-if Rails.env.development?
+if Rails.env.development? and !ENV['DO_NOT_FLUSH_REDIS']
   puts "Flushing redis (development mode)"
   $redis.flushall
 end
